# internal/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a chevron remote-signer server",
        "title": "Remote Signer API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://tldrlegal.com/license/mit-license"
        },
        "version": "1.4"
    },
    "basePath": "/remoteSigner",
    "paths": {
...
        "/sks/addKey": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Public Key Server",
                    "Key Store"
                ],
                "summary": "Adds a public key to the server",
                "operationId": "pks-add-public-key",
                "parameters": [
                    {
                        "description": "Information of the public key",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SKSAddKey"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
...
        "/sks/getKey": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Key Ring"
                ],
                "summary": "Fetches a public key from the server",
                "operationId": "kre-get-key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fingerprint of the key you want to fetch",
                        "name": "fingerPrint",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "gpg public key",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
...
        "/sks/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Key Server",
                    "Key Store"
                ],
                "summary": "Search for GPG keys by using it all fields.",
                "operationId": "pks-search-by-value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Value of the key you want to fetch",
                        "name": "valueData",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pagination Start Index (default: 0)",
                        "name": "pageStart",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination End Index (default: 100)",
                        "name": "pageEnd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GPGKey"
                            }
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
        "/sks/searchByEmail": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Key Server",
                    "Key Store"
                ],
                "summary": "Search for GPG keys by using it's email field.",
                "operationId": "pks-search-by-email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email of the key you want to fetch",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pagination Start Index (default: 0)",
                        "name": "pageStart",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination End Index (default: 100)",
                        "name": "pageEnd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GPGKey"
                            }
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
        "/sks/searchByFingerPrint": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Key Server",
                    "Key Store"
                ],
                "summary": "Search for GPG keys by using it's fingerprint field.",
                "operationId": "pks-search-by-fingerprint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fingerprint of the key you want to fetch",
                        "name": "fingerPrint",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pagination Start Index (default: 0)",
                        "name": "pageStart",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination End Index (default: 100)",
                        "name": "pageEnd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GPGKey"
                            }
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
        "/sks/searchByName": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Key Server",
                    "Key Store"
                ],
                "summary": "Search for GPG keys by using it's identifier name field.",
                "operationId": "pks-search-by-name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the key you want to fetch",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pagination Start Index (default: 0)",
                        "name": "pageStart",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination End Index (default: 100)",
                        "name": "pageEnd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GPGKey"
                            }
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "QuantoError.ErrorObject": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string",
                    "example": "INVALID_FIELD_DATA"
                },
                "errorData": {
                    "type": "object"
                },
                "errorField": {
                    "type": "string",
                    "example": "fieldName"
                },
                "message": {
                    "type": "string",
                    "example": "error example"
                },
                "stackTrace": {
                    "type": "string"
                }
            }
        },
...
        "models.GPGKey": {
            "type": "object",
            "properties": {
                "asciiArmoredPrivateKey": {
                    "type": "string"
                },
                "asciiArmoredPublicKey": {
                    "type": "string",
                    "example": "-----BEGIN PGP PUBLIC KEY BLOCK-----\nVersion: GnuPG v2\nComment: Generated by Chevron\n\nxsDNBF866vQBDADFS0xTbVzEgvGi6ZklWxuNHdO6ajtodno6XgtnAX74lHCuYTUk\nKX1E6AASdXOVrw4QQV8MUI2KFs6r6UhxKcRpw3M7SGeIYkyt5uWjZBYQFnCu3V8V\nfOAdXqbPplliZhfH2UbDOaWC97J4/8kOW8iAmFEL3DpvYF7N/wFx9VkR6T8qnOhV\njKsOmyh8CcxSQ0poxKtIcCpAfpTdG2fI2maux71kI8B3Fdu/fc/3GvTvy37giz9I\n9GHGEzbrWE2FoZeF4cUJC9ZiY6/zmPcTUIhe7HGjKcEjyZ+tqQ3cvJ1lVKXvhoJp\n0+nhY4nvFVQe0jNod/duJVDGxVBPDmQIPvD5FZAtQUgX0xb5Td4s/viw/7M5XjET\nWVeg6mvxn6Xaj6oo4kQiDF+00uOfqBljXxlxFMvH2NPnmx9H7XZ7/MXWl+YotUfT\nzW0VezpE8B9gkzZVir84icb5Of38DEqUovjJVGw9pEWd+intWeEKDuXt+iv3jXbP\n31hLqMFSr4Q9/U8AEQEAAc0hUmVtb3RlIFNpZ25lciBUZXN0IDx0ZXN0QHF1YW4u\ndG8+wsEUBBMBCgA+FiEEmF9o375LjCBYKXIwBVH0UqvkY6QFAl866vQCGwMFCQPC\nZwAFCwkIBwIGFQoJCAsCBBYCAwECHgECF4AACgkQBVH0UqvkY6Qjnwv/QzU1Qq0q\n1qffvy4l6NmpQXyI6AnIO5iG97SvDwtyxdkXVmCZM52p7V4nC3IPTaKP4r2OKH3D\n1UH+T11xwgucEw67aTte7mhkODyoBJ6mNj7bYZQx5SVQYL8dWQ5JvrS4ErXchW3j\n9sYyJMqSHEzizEXtvwRVun19DMWUYdrm3flaG5o5Fvr3OxG8/N1CuLe/R7HyhnwA\noP7VRQxAz9Ln6nBjpDRK1AdZ47ZsQflkRUl3boh6pLJ4UKIg3UHSLwfie1LSBKtj\n73X+LpLvuOQHuNa14KrWTiYAsdOmRPi/9lOg8O/t6oOMngzf4VAY+tiRgCTtIqjP\nIF9+G1rEdT6oY1j7eZhAXw0om8AH5V4TuSIRcFikRyAAWrYP8DA015lGSaORJAit\n2GULUKSZszV03m3o0SR55engvjR7CRuWmTbXdH8Eb5lGDJssUPCiPtGK1Y0v5DKb\n8lV0pMa3LqR0XT6bRmgtnqDM7FB0GE7AyIz721ikEKqiY3AXMmObOo7gzsDNBF86\n6vQBDAC/myEliUXeGP5TSGW5Et4p3DkAGK76G+o43Okyv5a8zEyEhXKaeEswGHqx\nan+6wz0iIqCE3xu54Gjaugb9dnCGmq4fD2Oly3nzkuC0eVE8dA0nYVuKFQZUpKwi\nEq7+UCMkndShKYcVTvcQk58sgQfZYkXtXjmklc/eeopA+zpoLmSnYe9ZGwrzR0Yn\n9qZkPWZ8OJNrbmtB9nsKNdmxkP8gzWAYzh5MGcd15FRQwpj6XDqMRkdQXu8Yo3Zq\nFQ/zZV4D9KlpQ/sqprYSGms1nmWIVExD5zCRSUmikUSJvVeSlMAkMDEufJMSpNY8\nxyeo6wu8vNPpKUINd8ZBcAWjyMkK8XUQKtd2cTafV1HWFeae/09NkiZsfjthKOVC\nMOIZMWssTUiu7NubznbMeFgVceuE4E1n9YHe5PtI76ybL0SqLIO93dvOD+yjHoWp\ndCFq3cAS7OXz24HHtBYzS9wkj+joJhFPJSo7WD1u6l/bJSm0g8gHUNIodHEpKo4P\np6BAeHUAEQEAAcLA/AQYAQoAJhYhBJhfaN++S4wgWClyMAVR9FKr5GOkBQJfOur0\nAhsMBQkDwmcAAAoJEAVR9FKr5GOkd08MAJLmpHHF8SE2kXRfY0/3imC0lHoJj5VP\na7OZEFPm9skBzECE3cinB4crCDdhLGJEhSYnbfnq/auf7dBtZS+QjulyGHjxNDfc\nitu8zxuq12phsyXZIMgjX5Cl1V1VGH3pnVm/nuSvwZ7Urew1pJ4Ep+xtRZhcwQcC\njYT29zPpIU2oLt50LDMdNmtUYmod1N23Tcd496GKevF/a01eZ3UA779jCvC8DS1s\nWH2DTx7aWUqi8gWa4xOZsBJlyypLZDpDPETp2/+WFllWM96ubyApvkwZIOGggnwM\nQXbJ32m5vVxgQkUYl98VFEttka3rTQtP+Hnfntqj2LVl54VKUhBiGRPOC8OrpAo5\nHY/Jk2dZafMtTlbiQdgzCw3LB9n4Mc7V7d7rJT7DWq1G09lAlQWk/3r2JmBbayGp\n9UlipL+H4r4AOQirwmuaHMJ9bHCnzgAUMHomw0NDktkDnnPKZ2TxcSD9m4qgrf1q\nFXkbQHgIocl4wcuq7ZegIr7Z7hYVd0EfOA==\n=Ufu+\n-----END PGP PUBLIC KEY BLOCK-----"
                },
                "emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "test@quan.to"
                    ]
                },
                "fullFingerprint": {
                    "type": "string",
                    "example": "0551F452ABE463A4"
                },
                "id": {
                    "type": "string"
                },
                "keyBits": {
                    "type": "integer",
                    "example": 3072
                },
                "keyUids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GPGKeyUid"
                    }
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Remote Signer Test"
                    ]
                },
                "parentKey": {
                    "$ref": "#/definitions/models.GPGKey"
                },
                "subkeys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.GPGKeyUid": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "test@quan.to"
                },
                "name": {
                    "type": "string",
                    "example": "Remote Signer Test"
                }
            }
        },
...
        "models.SKSAddKey": {
            "type": "object",
            "properties": {
                "publicKey": {
                    "type": "string",
                    "example": "-----BEGIN PGP PUBLIC KEY BLOCK-----\nVersion: GnuPG v2\nComment: Generated by Chevron\n\nxsDNBF866vQBDADFS0xTbVzEgvGi6ZklWxuNHdO6ajtodno6XgtnAX74lHCuYTUk\nKX1E6AASdXOVrw4QQV8MUI2KFs6r6UhxKcRpw3M7SGeIYkyt5uWjZBYQFnCu3V8V\nfOAdXqbPplliZhfH2UbDOaWC97J4/8kOW8iAmFEL3DpvYF7N/wFx9VkR6T8qnOhV\njKsOmyh8CcxSQ0poxKtIcCpAfpTdG2fI2maux71kI8B3Fdu/fc/3GvTvy37giz9I\n9GHGEzbrWE2FoZeF4cUJC9ZiY6/zmPcTUIhe7HGjKcEjyZ+tqQ3cvJ1lVKXvhoJp\n0+nhY4nvFVQe0jNod/duJVDGxVBPDmQIPvD5FZAtQUgX0xb5Td4s/viw/7M5XjET\nWVeg6mvxn6Xaj6oo4kQiDF+00uOfqBljXxlxFMvH2NPnmx9H7XZ7/MXWl+YotUfT\nzW0VezpE8B9gkzZVir84icb5Of38DEqUovjJVGw9pEWd+intWeEKDuXt+iv3jXbP\n31hLqMFSr4Q9/U8AEQEAAc0hUmVtb3RlIFNpZ25lciBUZXN0IDx0ZXN0QHF1YW4u\ndG8+wsEUBBMBCgA+FiEEmF9o375LjCBYKXIwBVH0UqvkY6QFAl866vQCGwMFCQPC\nZwAFCwkIBwIGFQoJCAsCBBYCAwECHgECF4AACgkQBVH0UqvkY6Qjnwv/QzU1Qq0q\n1qffvy4l6NmpQXyI6AnIO5iG97SvDwtyxdkXVmCZM52p7V4nC3IPTaKP4r2OKH3D\n1UH+T11xwgucEw67aTte7mhkODyoBJ6mNj7bYZQx5SVQYL8dWQ5JvrS4ErXchW3j\n9sYyJMqSHEzizEXtvwRVun19DMWUYdrm3flaG5o5Fvr3OxG8/N1CuLe/R7HyhnwA\noP7VRQxAz9Ln6nBjpDRK1AdZ47ZsQflkRUl3boh6pLJ4UKIg3UHSLwfie1LSBKtj\n73X+LpLvuOQHuNa14KrWTiYAsdOmRPi/9lOg8O/t6oOMngzf4VAY+tiRgCTtIqjP\nIF9+G1rEdT6oY1j7eZhAXw0om8AH5V4TuSIRcFikRyAAWrYP8DA015lGSaORJAit\n2GULUKSZszV03m3o0SR55engvjR7CRuWmTbXdH8Eb5lGDJssUPCiPtGK1Y0v5DKb\n8lV0pMa3LqR0XT6bRmgtnqDM7FB0GE7AyIz721ikEKqiY3AXMmObOo7gzsDNBF86\n6vQBDAC/myEliUXeGP5TSGW5Et4p3DkAGK76G+o43Okyv5a8zEyEhXKaeEswGHqx\nan+6wz0iIqCE3xu54Gjaugb9dnCGmq4fD2Oly3nzkuC0eVE8dA0nYVuKFQZUpKwi\nEq7+UCMkndShKYcVTvcQk58sgQfZYkXtXjmklc/eeopA+zpoLmSnYe9ZGwrzR0Yn\n9qZkPWZ8OJNrbmtB9nsKNdmxkP8gzWAYzh5MGcd15FRQwpj6XDqMRkdQXu8Yo3Zq\nFQ/zZV4D9KlpQ/sqprYSGms1nmWIVExD5zCRSUmikUSJvVeSlMAkMDEufJMSpNY8\nxyeo6wu8vNPpKUINd8ZBcAWjyMkK8XUQKtd2cTafV1HWFeae/09NkiZsfjthKOVC\nMOIZMWssTUiu7NubznbMeFgVceuE4E1n9YHe5PtI76ybL0SqLIO93dvOD+yjHoWp\ndCFq3cAS7OXz24HHtBYzS9wkj+joJhFPJSo7WD1u6l/bJSm0g8gHUNIodHEpKo4P\np6BAeHUAEQEAAcLA/AQYAQoAJhYhBJhfaN++S4wgWClyMAVR9FKr5GOkBQJfOur0\nAhsMBQkDwmcAAAoJEAVR9FKr5GOkd08MAJLmpHHF8SE2kXRfY0/3imC0lHoJj5VP\na7OZEFPm9skBzECE3cinB4crCDdhLGJEhSYnbfnq/auf7dBtZS+QjulyGHjxNDfc\nitu8zxuq12phsyXZIMgjX5Cl1V1VGH3pnVm/nuSvwZ7Urew1pJ4Ep+xtRZhcwQcC\njYT29zPpIU2oLt50LDMdNmtUYmod1N23Tcd496GKevF/a01eZ3UA779jCvC8DS1s\nWH2DTx7aWUqi8gWa4xOZsBJlyypLZDpDPETp2/+WFllWM96ubyApvkwZIOGggnwM\nQXbJ32m5vVxgQkUYl98VFEttka3rTQtP+Hnfntqj2LVl54VKUhBiGRPOC8OrpAo5\nHY/Jk2dZafMtTlbiQdgzCw3LB9n4Mc7V7d7rJT7DWq1G09lAlQWk/3r2JmBbayGp\n9UlipL+H4r4AOQirwmuaHMJ9bHCnzgAUMHomw0NDktkDnnPKZ2TxcSD9m4qgrf1q\nFXkbQHgIocl4wcuq7ZegIr7Z7hYVd0EfOA==\n=Ufu+\n-----END PGP PUBLIC KEY BLOCK-----"
                }
            }
        },
...
    },
...
}